pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "exca-dance"
version = "0.1.0"
description = "Excavator rhythm game for excavator training"
requires-python = ">=3.10"
dependencies = [
    "pygame-ce>=2.4",
    "moderngl>=5.10",
    "PyOpenGL>=3.1",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1"]
ros2 = []  # rclpy and sensor_msgs installed separately via ROS2

[project.scripts]
exca-dance = "exca_dance.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.ruff]
line-length = 100
target-version = "py310"
